requirements.txt
fastapi>=0.115.0
python-dotenv==1.0.1
google-generativeai==0.8.3
uvicorn[standard]>=0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.35
passlib==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
crewai[google-genai]>=1.14.0
